pyproject.toml
[project]
name = "peerdas"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-connect>=0.8.0",
    "pandas>=2.0",
    "plotly>=5.0",
    "altair>=5.0",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "pyarrow>=22.0.0",
    "jupytext>=1.18.1",
    "jupyterlab>=4.5.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "nbconvert>=7.16.6",
    "papermill>=2.6.0",
    "jupyter-book>=1.0.0",
]
